+++ app/Imports/ProductReplacementsImport.php
@@ -4,6 +4,7 @@ namespace App\Imports;
 
 use App\Http\Controllers\Support\ProductReplacementController;
 use App\Product;
+use Illuminate\Support\Facades\Log;
 use Maatwebsite\Excel\Concerns\Importable;
 use Maatwebsite\Excel\Concerns\OnEachRow;
 use Maatwebsite\Excel\Concerns\WithChunkReading;
@@ -28,31 +29,20 @@ class ProductReplacementsImport implements WithStartRow, OnEachRow, WithChunkRea
 
     public function onRow(Row $row)
     {
+        // TODO Get headers from the template file to make the app more generic
         $itemFromFile = 'item';
         $replacementFromFile = 'replacement';
         $remarkFromFile = 'remark';
 
         $row = $row->toArray();
 
-        // Check columns
-        if (is_null($row[$itemFromFile] ?? null)) {
-            //! Return with error message, fix!
-            return redirect()->action([ProductReplacementController::class, 'importReplacementForm'])->with('danger', 'Kolumnen ' . $item . ' saknas');
-        }
-        if (is_null($row[$replacementFromFile] ?? null)) {
-            //! Return with error message, fix!
-            return;
-        }
-        if (is_null($row[$remarkFromFile] ?? null)) {
-            //! Return with error message, fix!
-            return;
-        }
 
         // Update comment if record exist, or create
         $item = Product::where('item', $row[$itemFromFile])->first();
+        // TODO Check if product doesn't exist, insert it to products table and notify that it should be updated with more data
+        
         $replacement = Product::where('item', $row[$replacementFromFile])->first();
-        // If product doesn't exist, insert it to products table and notify that it should be updated with more data
-        // Check if replacement product exist in Products table, otherwise insert..
+        // TODO Check if replacement product exist in Products table, otherwise insert..
 
         // Check if replacement exist in replacement-table
         if ($item->replacements->contains($replacement)) {
@@ -60,8 +50,6 @@ class ProductReplacementsImport implements WithStartRow, OnEachRow, WithChunkRea
             
         } else {
             $item->replacements()->attach($replacement,['comment' => $row[$remarkFromFile]]);
-            // $item->replacements()->updateExistingPivot($item, ['comment' => $row[$remarkFromFile]]);
-            // dd ($item->item . ' tillagd');
         }
     }        
 }